refactor: types에 있던 post 관련 type들 분리
